fix(npm): add engine-only core entry for Node ESM
