+++ npm/package.json
@@ -11,6 +11,10 @@
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
     },
+    "./core": {
+      "types": "./dist/core.d.ts",
+      "import": "./dist/core.js"
+    },
     "./react": {
       "types": "./dist/react.d.ts",
       "import": "./dist/react.js"
@@ -32,6 +36,9 @@
   },
   "typesVersions": {
     "*": {
+      "core": [
+        "./dist/core.d.ts"
+      ],
       "react": [
         "./dist/react.d.ts"
       ],
